requirements.txt
absl-py
Cython
jax
matplotlib
seaborn
ml_collections
numpy
ogb
polars
rdkit
scipy
torch_geometric
torchmetrics
wandb
lightning
dvc-gs
ConfigSpace
hpbandster
ray[train]
ray[tune]
wheel
teqp
PCSAFTsuperanc
feos
si-units
xgboost
onnx
onnxscript
onnxruntime